requirements.txt
# PyTorch with CUDA 12.1 support
torch==2.3.0+cu121
torchvision --extra-index-url https://download.pytorch.org/whl/cu121

# Computer Vision
opencv-python-headless==4.8.1.78
pillow==10.0.1

# AI/ML
transformers==4.36.0
accelerate==0.24.1
qwen-vl-utils==0.0.1

# Data Processing
pandas==2.1.3
openpyxl==3.1.2

# Additional utilities
numpy==1.24.3 